Move test data files to their own directory
