+++ EconDataLens.Tests/TestData/cu.area.empty
