+++ EconDataLens.Tests/TestData/cu.area.sample
