+++ EconDataLens.Tests/TestData/cu.data.0.Current.sample
