+++ EconDataLens.Tests/TestData/cu.period.sample
