+++ EconDataLens.Tests/TestData/cu.series.empty
