+++ EconDataLens.Tests/TestData/cu.series.sample
